First commit
I setup the template for our webpage and added the basic input fields
for the login process. I made a shitty ass logo and figured out how to
arrange the service bar so it is fixed at the top of the page. I also
created the php file for the service bar. This should contain the
scripts for getting the values of the input fields and checking them
with the user's accounts in the database.