Validation (not complete, needs testing/advice)
index.php - Added name and id attr. to login form.
index.php - removed method attr in registration (Don't know why I put it
there.)
validation.js - Added the remote to the rules. It seems to connect to
the database fine with no errors, but could use more testing.
validation.js - I ran into a problem for log in validation. If you enter
a username then remove it, the username error shows. If you add a
password, then remove the password, the error for the password does not
show up. I just need a test and advice on a fix.
validation.js - I tried the equalTo in the registration rules to verify
passwords, but doesn't seem to work. Not sure what is wrong there
either.